requirements.txt
rtyaml
gitpython
